+++ contextMenus/end.js
@@ -1,40 +1,41 @@
 import { ContextMenuCommandBuilder, ApplicationCommandType } from "discord.js"
-import { db, end } from "helpers"
+import { db } from "../helpers/database.js"
+import { end } from "../helpers/end.js"
 
 export default {
-    data: new ContextMenuCommandBuilder()
-        .setName("End Giveaway")
-        .setType(ApplicationCommandType.Message),
-    execute: async (interaction) => {
-        const giveaway = await db.Giveaways.findOne({
-            where: {
-                messageId: interaction.targetId,
-            },
-        })
+	data: new ContextMenuCommandBuilder()
+		.setName("End Giveaway")
+		.setType(ApplicationCommandType.Message),
+	execute: async (interaction) => {
+		const giveaway = await db.Giveaways.findOne({
+			where: {
+				messageId: interaction.targetId
+			}
+		})
 
-        if (!giveaway)
-            return await interaction.reply({
-                content: "That message is not a giveaway.",
-                ephemeral: true,
-            })
+		if (!giveaway)
+			return await interaction.reply({
+				content: "That message is not a giveaway.",
+				ephemeral: true
+			})
 
-        if (interaction.user.id != giveaway.userId)
-            return await interaction.reply({
-                content: "You are not the giveaway hoster.",
-                ephemeral: true,
-            })
+		if (interaction.user.id != giveaway.userId)
+			return await interaction.reply({
+				content: "You are not the giveaway hoster.",
+				ephemeral: true
+			})
 
-        if (giveaway.isFinished)
-            return await interaction.reply({
-                content: "The giveaway has already ended.",
-                ephemeral: true,
-            })
+		if (giveaway.isFinished)
+			return await interaction.reply({
+				content: "The giveaway has already ended.",
+				ephemeral: true
+			})
 
-        end(giveaway, interaction.client, true)
+		end(giveaway, interaction.client, true)
 
-        await interaction.reply({
-            content: "Giveaway Ended.",
-            ephemeral: true,
-        })
-    },
+		await interaction.reply({
+			content: "Giveaway Ended.",
+			ephemeral: true
+		})
+	}
 }
