--- helpers/package.json
@@ -1,6 +1,0 @@
-{
-    "name": "helpers",
-    "version": "1.0.0",
-    "main": "index.js",
-    "type": "module"
-}
